requirements.txt
os
torch
torchvision
pandas
numpy
matplotlib.pyplot
sklearn
cv2
cvlib
